requirements.txt
streamlit
python-docx
mammoth
docx2pdf
pdf2image
Pillow
streamlit
python-docx
pandas
openpyxl
mammoth
